refactor: modularize Error
