+++ src/init.rs
@@ -11,8 +11,10 @@ use inflector::cases::snakecase::to_snake_case;
 use probe_rs::config::{get_target_by_name, search_chips};
 
 use crate::types::{
-    family::mem_region::MemRegion, family::Family, Chip, Error, InitArgs, InvalidChip,
-    PanicHandler, Softdevice, Target,
+    error::{Error, InvalidChip},
+    family::mem_region::MemRegion,
+    family::Family,
+    Chip, InitArgs, PanicHandler, Softdevice, Target,
 };
 
 pub struct Init {
